Creacion packages view y presenter Se creo la clase ArtistDescriptionFormatterHtml en la que se formatea la informacion del artista.
