+++ app/src/main/java/ayds/winchester/songinfo/moredetails/fulllogic/presentation/view/OtherInfoWindowUiState.kt
@@ -1,4 +1,4 @@
-package ayds.winchester.songinfo.moredetails.fulllogic.presentation
+package ayds.winchester.songinfo.moredetails.fulllogic.presentation.view
 
 private const val WIKIPEDIA_LOGO_URL = "https://upload.wikimedia.org/wikipedia/commons/8/8c/Wikipedia-logo-v2-es.png"
